+++ apps/web/layouts/main-layout/index.module.scss
@@ -95,6 +95,4 @@
 .content {
   width: 24rem;
   margin: 0.2rem auto;
-  border: 1px solid var(--theme-border-color);
-  overflow: hidden;
 }
