+++ apps/web/pages/_app.tsx
@@ -1,10 +1,12 @@
 import Head from "next/head";
 import { useEffect } from "react";
-import { AppProps } from "../type";
-import "./global.css";
+import { AppProps } from "../types";
+
+import "../styles/global.css";
 import "@arco-design/web-react/dist/css/arco.css";
-import "./normalize.css";
-import "./common.css";
+import "../styles/normalize.css";
+import "../styles/common.css";
+
 export default function App(props: AppProps) {
   const { Component } = props;
   useEffect(() => {
@@ -12,11 +14,11 @@ export default function App(props: AppProps) {
   }, []);
   return (
     <>
-      <Component></Component>
       <Head>
         <title>Code Show</title>
         <link rel="icon" href="/favicon.ico" />
       </Head>
+      <Component />
     </>
   );
 }
