+++ apps/web/pages/detail/index.tsx
@@ -5,7 +5,7 @@ import "highlight.js/styles/night-owl.css";
 import { useEffect, useState } from "react";
 import React, { FC } from "react";
 import { getArticleDetail, getComment } from "api-sdk";
-import { Code, Comment } from "../../type";
+import { Code, Comment } from "../../types";
 import MyComment from "../../components/comment";
 import { useRouter } from "next/router";
 import MyCode from "../../components/code";
