--- apps/web/pages/index.module.css
@@ -1,96 +1,0 @@
-.container {
-  width: 100%;
-  min-height: 100vh;
-  position: relative;
-  background-color: #24273b;
-}
-
-.head {
-  display: flex;
-  justify-content: space-between;
-  height: 1.2rem;
-  line-height: 1.2rem;
-  background-color: var(--theme-bg-color);
-  color: var(--theme-light-text-color);
-  grid-area: header;
-  position: fixed;
-  top: 0;
-  right: 0;
-  width: 100%;
-  z-index: 111;
-}
-
-.head_l {
-  caret-color: transparent;
-  margin-left: 0.4rem;
-  font-family: "Telefon Black", Sans-Serif;
-  font-size: 0.4rem;
-}
-.head_r {
-  margin-right: 0.4rem;
-  display: flex;
-  justify-content: flex-start;
-}
-.search {
-  margin-top: -0.04rem;
-  background-color: #151728;
-  margin-right: 0.4rem;
-  position: relative;
-}
-.search_input {
-  padding-left: 0.16rem;
-  height: 0.64rem;
-  border-radius: 0.1rem;
-  border: none;
-  background-color: #2c303a;
-  color: #aaaebc;
-  cursor: pointer;
-  transition: 0.36s ease;
-}
-.search_input:focus {
-  outline: none;
-  background-color: #444857;
-  cursor: pointer;
-  transition: 0.36s ease;
-}
-.icon_search {
-  color: #707480;
-  position: absolute;
-  right: 0.16rem;
-  top: 0.44rem;
-  font-size: 0.34rem;
-}
-.publish {
-  font-size: 0.28rem;
-  height: 0.64rem;
-  margin-right: 0.36rem;
-  background-color: #444857;
-  color: var(--theme-light-text-color);
-  margin-top: 0.24rem;
-  border-radius: 0.1rem;
-}
-.publish:hover {
-  background-color: #5a5f73;
-}
-.login {
-  font-size: 0.28rem;
-  height: 0.64rem;
-  margin-right: 0.4rem;
-  background-color: #444857;
-  color: var(--theme-light-text-color);
-  margin-top: 0.24rem;
-  border-radius: 0.1rem;
-}
-.login:hover {
-  background-color: #5a5f73;
-}
-/* 主体部分 */
-.content {
-  width: calc(100% - 5.5rem);
-  margin: 0rem auto;
-  padding: 1.8rem 0 0 0;
-  display: flex;
-  justify-content: space-between;
-  height: calc(100vh-60px);
-  caret-color: transparent;
-}
