+++ apps/web/pages/personal/index.module.scss
@@ -1,2 +1,17 @@
 .personal {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+}
+
+.left {
+  width: 18.5rem;
+  border: 1px solid var(--theme-border-color);
+}
+
+.right {
+  width: 5rem;
+  border: 1px solid var(--theme-border-color);
 }
